refactor(ors_control_app): extract routes/status and routes/services
Extract status + services route groups from server/index.ts to dedicated
Express Router modules:
- routes/status.ts: /api/status, /api/config, /api/health, /api/ors-readiness
  (including the StageProbe / probeStage / phasesFor / buildReadiness helpers
  that were embedded inline in /api/ors-readiness)
- routes/services.ts: /api/resume, /api/suspend,
  /api/services/:name/(resume|suspend), /api/scale

server/index.ts mounts them via app.use(createStatusRouter(APP_VERSION))
and app.use(createServicesRouter()). Lines down 3,428 -> 3,160.

Phase 1 of holistic-control-app-refactor-final, route extraction step 1.
No behaviour change. tsc -p tsconfig.server.json passes cleanly.
